Increased wait duration of GitLab running status.

+++ src/run_ci_job.sh
@@ -6,16 +6,20 @@ source src/creds.txt
 source src/create_personal_access_token.sh
 
 # TODO: change 127.0.0.1 with gitlab server address variable
+# TODO: ensure the receipe works every time, instead of every other time.
+# There currently is an error when the gitlab repo is deleted or cloned, which is
+# resolved the second time the function is called because at that time the repo is
+# deleted or cloned/created.
 
 #source src/run_ci_job.sh && receipe
 receipe() {
-	$(delete_target_folder)
-	$(delete_repository)
-	$(create_repository)
-	$(clone_repository)
-	$(export_repo)
-	$(commit_changes)
-	$(push_changes)
+	delete_target_folder
+	delete_repository
+	create_repository
+	clone_repository
+	export_repo
+	commit_changes
+	push_changes
 }
 
 commit_changes() {
@@ -30,7 +34,7 @@ push_changes() {
 	gitlab_username=$(echo $gitlab_server_account | tr -d '\r')
 	gitlab_server_password=$(echo $gitlab_server_password | tr -d '\r')
 	output=$(cd ../$SOURCE_FOLDERNAME && git push http://$gitlab_username:$gitlab_server_password@127.0.0.1/$gitlab_username/$repo_name.git)
-	echo "outrput=$output"
+	echo "output=$output"
 }
 
 # source src/run_ci_job.sh && export_repo
@@ -39,7 +43,7 @@ delete_target_folder() {
 	# check if target folder already exists
 	# delete target folder if it already exists
 	if [ -d "../$SOURCE_FOLDERNAME" ] ; then
-	    rm -r "../$SOURCE_FOLDERNAME"
+	    sudo rm -r "../$SOURCE_FOLDERNAME"
 	fi
 	# create target folder
 	# copy source folder to target
@@ -61,15 +65,24 @@ export_repo() {
 create_repository() {
 #source src/run_ci_job.sh && create_repository
 	# Create personal GitLab access token
-	create_gitlab_personal_access_token
+	# TODO: re-enable
+	#create_gitlab_personal_access_token
 	
 	# load personal_access_token
 	personal_access_token=$(echo $GITLAB_PERSONAL_ACCESS_TOKEN | tr -d '\r')
 	
 	# Create repo named foobar
 	repo_name=$SOURCE_FOLDERNAME
-	output=$(curl -H "Content-Type:application/json" http://127.0.0.1/api/v4/projects?private_token=$personal_access_token -d "{ \"name\": \"$repo_name\" }")
-	echo "output=$output"
+	{ # try
+		output=$(curl -H "Content-Type:application/json" http://127.0.0.1/api/v4/projects?private_token=$personal_access_token -d "{ \"name\": \"$repo_name\" }")
+		echo "output=$output"
+		#save your output
+		true
+	} || { # catch
+		# save log for exception 
+		true
+	}
+	
 }
 
 #source src/run_ci_job.sh && delete_repository
@@ -77,12 +90,24 @@ delete_repository() {
 	# load personal_access_token
 	personal_access_token=$(echo $GITLAB_PERSONAL_ACCESS_TOKEN | tr -d '\r')
 	
-	# Create repo named foobar
 	gitlab_username=$(echo $gitlab_server_account | tr -d '\r')
+	gitlab_server_password=$(echo $gitlab_server_password | tr -d '\r')
 	repo_name=$SOURCE_FOLDERNAME
 	
-	output=$(curl -H 'Content-Type: application/json' -H "Private-Token: $personal_access_token" -X DELETE http://127.0.0.1/api/v4/projects/$gitlab_username%2F$repo_name)
-	echo "output=$output"
+	# TODO: check if the repo exists (unstable behaviour, sometimes empty when repository DOES exist).
+	exists=$(git ls-remote --exit-code -h "http://$gitlab_username:$gitlab_server_password@127.0.0.1/$gitlab_username/$repo_name")
+	echo "exists=$exists"
+	# DELETE the repository
+	if [ -z "$exists" ]; then
+		echo "Repo does not exist."
+	else
+		output=$(curl -H 'Content-Type: application/json' -H "Private-Token: $personal_access_token" -X DELETE http://127.0.0.1/api/v4/projects/$gitlab_username%2F$repo_name)
+	fi
+	
+	# TODO: loop untill repository is deleted (otherwise the following error is thrown:
+	# TODO: check if the repo exists
+	#output={"message":{"base":["The project is still being deleted. Please try again later."],"limit_reached":[]}}
+
 }
 
 #source src/run_ci_job.sh && clone_repository
@@ -90,11 +115,10 @@ clone_repository() {
 	repo_name=$(echo $SOURCE_FOLDERNAME | tr -d '\r')
 	gitlab_username=$(echo $gitlab_server_account | tr -d '\r')
 	gitlab_server_password=$(echo $gitlab_server_password | tr -d '\r')
-	#git@127.0.0.1:root/foobar.git
-	#
-	rm -r ../$repo_name
-	echo "/$gitlab_server_account=$gitlab_server_account"
-	echo "/$gitlab_server_password=$gitlab_server_password"
+	
+	#sudo rm -r ../$repo_name
+	#echo "/$gitlab_server_account=$gitlab_server_account"
+	#echo "/$gitlab_server_password=$gitlab_server_password"
 	output=$(cd .. && git clone http://$gitlab_username:$gitlab_server_password@127.0.0.1/$gitlab_username/$repo_name.git)
 	echo "output=$output"
 }
